requirements.txt
pandas
numpy
pymongo
scikit-learn
google-generativeai
tqdm
python-dotenv
bson
aiohttp
